# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server File store server"
  version: "1.0.0"
  title: "Filestore"
schemes:
- "http"
host: localhost:8080
basePath: "/"
consumes:
 - application/json
produces:
 - application/json
paths:
  /file:
    get:
      summary: "Find by File Name"
      description: "Returns a single File Name"
      operationId: "readFile"
      produces:
      - "application/json"
      parameters:
      - name: "fileName"
        in: "query"
        description: "Name of the file to return"
        type: "string"
        required: true
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/FileDetailsSO"
        400:
          description: "Invalid File Name supplied"
        404:
          description: "File not found"
    delete:
      summary: "Deletes a file "
      operationId: "deleteFile"
      produces:
      - "application/json"
      parameters:
      - name: "fileName"
        in: "query"
        description: "File Name to delete"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
            items:
              $ref: "#/definitions/Message"
        400:
          description: "Invalid File Name supplied"
        404:
          description: "File not found"
  /file/list:
    get:
      summary: "Finds files in the Folder Path"
      description: "Display the file name"
      operationId: "fileNamesList"
      produces:
      - "application/json"
      parameters:
      - name: "folderPath"
        in: "query"
        description: "Name of the file to return"
        type: "string"
        required: true
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/FileNames"
        400:
          description: "Invalid File Path"
  /file:
    post:
      summary: "uploads a File"
      description: ""
      operationId: "uploadFile"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "additionalMetadata"
        in: "formData"
        description: "Additional data to pass to server"
        required: false
        type: "string"
      - name: "file"
        in: "formData"
        description: "file to upload"
        required: false
        type: "file"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "object"
            items:
              $ref: "#/definitions/InputStreamResource"
definitions:
  FileDetailsSO:
    type: "object"
    properties:
      fileName:
        type: "string"
      createdDate:
        type: "string"
  FileSO:
    type: "object"
    properties:
      fileDetails:
        $ref: "#/definitions/FileDetailsSO"
  FileNames:
        type: "array"
        items:
          type: "string"
  Message:
        type: "string"
        items:
          type: "string"
  InputStreamResource:
        type: "object"
        items:
          type: "object"